make Factorization backend agnostic
