+++ funfact/model/_factorization.py
@@ -12,10 +12,9 @@ from funfact.lang.interpreter import (
     ShapeAnalyzer,
     Vectorizer
 )
-from jax.tree_util import register_pytree_node_class
+from funfact.backend import active_backend as ab
 
 
-@register_pytree_node_class
 class Factorization:
     '''A factorization model is a concrete realization of a tensor expression.
     The factor tensors of the model can be optimized to obtain a better
@@ -197,11 +196,7 @@ class Factorization:
         ):
             n.data = tensors[i]
 
-    def tree_flatten(self):
-        return list(self.factors), self.tsrex
 
-    @classmethod
-    def tree_unflatten(cls, tsrex, tensors):
-        unflatten = cls(tsrex, initialize=False)
-        unflatten.factors = tensors
-        return unflatten
+@ab.autograd_decorator
+class AutoGradFactorization(Factorization, ab.AutoGradMixin):
+    pass
